Latest Version
I fixed a few of the grid layouts for the edit flashcard; however, it's not fully completed. So what I have to do is to fix the backend side where the users will have to be able to add cards inside the file without having a base. The layout will need a bit of fixing too. Should be done soon.