Change minds
Alright this is a long one but I changed my mind from having different features on a different files. I decided to change back to the idea of having all in one file to make it easy to access, and don't need to import from multiple files. Furthermore, I also added the shop frame into the UI, although the UI is still changing sizes everytime I change frames, it's something I need to work on. The last thing I need to work on would be the functionality of the powerups. Since I only have the ability to buy and use, I still haven't added the backend side of the usage yet. So functionality and stability to hold the frames together.