Bug Fixes
A lot has changed. I fixed the GUI where the home page is now a lot more organized. Then in the flashcard page, I added a lot of apply_theme functions to different buttons when they are clicked. The work is due Sunday 2nd November which is today. So I'm grinding through the morning.